+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.2.1
 
 * BUGFIX - Update CLI tool to not call server/server.js directly
+* FEATURE - Update CLI tool to provide more options e.g limit number of scanning folders, skip existing records.
 
 ## v0.2.0